Change the title of the application shell
The application shell is the page displayed when the application is
launched. It is controlled by an Angular component named AppComponent.
The implementation of the shell AppComponent is distributed over three
files:
- app.component.ts: The component class code, written in TypeScript.
- app.component.html: The component template, written in HTML.
- app.component.css: The component's private CSS styles.
To change the title of the application shell, set the value of the title
property of the AppComponent class.